… src/System.Runtime.Numerics/src/System/Numerics/BigIntegerCalculator.DivRem.cs
… src/System.Runtime.Numerics/src/System/Numerics/BigIntegerCalculator.SquMul.cs