To introduce further performance tweaks, some minor improvements are added to `BigIntegerCalculator`, which cover all the basic operations. Having these algorithms based on raw uint[] objects will lead to more interesting tweaks of the more fancy stuff (`ModPow`).

A basic performance comparison based on [this code](https://gist.github.com/axelheer/3b701c91271b7c762586) unveils the following results:

**Add**

| # of bits | # of vals | before ms | after ms |
|----------:|----------:|----------:|---------:|
|        16 | 1,000,000 |        40 |       37 |
|        64 | 1,000,000 |        97 |       94 |
|       256 | 1,000,000 |       110 |      107 |
|     1,024 | 1,000,000 |       163 |      152 |
|     4,096 |   100,000 |        40 |       37 |
|    16,384 |   100,000 |       147 |      122 |
|    65,536 |   100,000 |       515 |      440 |

**Subtract**

| # of bits | # of vals | before ms | after ms |
|----------:|----------:|----------:|---------:|
|        16 | 1,000,000 |        44 |       38 |
|        64 | 1,000,000 |       106 |       95 |
|       256 | 1,000,000 |       128 |      111 |
|     1,024 | 1,000,000 |       182 |      164 |
|     4,096 |   100,000 |        42 |       38 |
|    16,384 |   100,000 |       143 |      131 |
|    65,536 |   100,000 |       511 |      444 |

**Divide**

| # of bits | # of vals | before ms | after ms |
|----------:|----------:|----------:|---------:|
|        16 | 1,000,000 |        49 |       29 |
|        64 | 1,000,000 |       125 |      110 |
|       256 | 1,000,000 |       349 |      287 |
|     1,024 |   100,000 |       215 |      119 |
|     4,096 |    10,000 |       254 |      124 |
|    16,384 |     1,000 |       378 |      172 |
|    65,536 |       100 |       595 |      263 |

*Note:* during making this pull request I noted that all the `Debug.Assert` aren't doing anything... (?)

*Attention:* because of using `stackalloc` I had to reduce an unit test value, which is currently very high and leads to a stack overflow. Would it be better to switch to `fixed` or should the actual stack size be increased, if someone really needs that huge numbers?

#1307