@stephentoub 2015-05-04
As can be seen by looking in the CI log, this change is causing multiple asserts when run on Linux:
```
13:53:27 [39;49m[31m   System.Numerics.Tests.BigIntegerConstructorTest.RunCtorInt64Tests [FAIL]
13:53:28             at System.Diagnostics.Debug.Assert(Boolean condition)
13:53:28             at System.Numerics.BigIntegerCalculator.LeftShift(UInt32* value, Int32 valueLength, UInt32* target, Int32 targetLength, Int32 shift)
13:53:28             at System.Numerics.BigIntegerCalculator.Divide(UInt32* left, Int32 leftLength, UInt32* right, Int32 rightLength, UInt32* quotient, Int32 quotientLength, UInt32* remainder, Int32 remainderLength)
13:53:28             at System.Numerics.BigIntegerCalculator.Divide(UInt32[] left, UInt32[] right)
13:53:28             at System.Numerics.BigInteger.op_Division(BigInteger dividend, BigInteger divisor)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.VerifyBigIntegerUsingIdentities(BigInteger bigInteger, Boolean isZero)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.VerifyCtorInt64(Int64 value)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.RunCtorInt64Tests()
```
```
13:53:28 [31m   System.Numerics.Tests.BigIntegerConstructorTest.RunCtorUInt64Tests [FAIL]
13:53:28             at System.Diagnostics.Debug.Assert(Boolean condition)
13:53:28             at System.Numerics.BigIntegerCalculator.LeftShift(UInt32* value, Int32 valueLength, UInt32* target, Int32 targetLength, Int32 shift)
13:53:28             at System.Numerics.BigIntegerCalculator.Divide(UInt32* left, Int32 leftLength, UInt32* right, Int32 rightLength, UInt32* quotient, Int32 quotientLength, UInt32* remainder, Int32 remainderLength)
13:53:28             at System.Numerics.BigIntegerCalculator.Divide(UInt32[] left, UInt32[] right)
13:53:28             at System.Numerics.BigInteger.op_Division(BigInteger dividend, BigInteger divisor)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.VerifyBigIntegerUsingIdentities(BigInteger bigInteger, Boolean isZero)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.VerifyCtorUInt64(UInt64 value)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.RunCtorUInt64Tests()
```
```
13:53:28 [31m   System.Numerics.Tests.BigIntegerConstructorTest.RunCtorSingleTests [FAIL]
13:53:28             at System.Diagnostics.Debug.Assert(Boolean condition)
13:53:28             at System.Numerics.BigIntegerCalculator.LeftShift(UInt32* value, Int32 valueLength, UInt32* target, Int32 targetLength, Int32 shift)
13:53:28             at System.Numerics.BigIntegerCalculator.Divide(UInt32* left, Int32 leftLength, UInt32* right, Int32 rightLength, UInt32* quotient, Int32 quotientLength, UInt32* remainder, Int32 remainderLength)
13:53:28             at System.Numerics.BigIntegerCalculator.Divide(UInt32[] left, UInt32[] right)
13:53:28             at System.Numerics.BigInteger.op_Division(BigInteger dividend, BigInteger divisor)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.VerifyBigIntegerUsingIdentities(BigInteger bigInteger, Boolean isZero)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.VerifyCtorSingle(Single value)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.RunCtorSingleTests()
```
```
13:53:28 [31m   System.Numerics.Tests.BigIntegerConstructorTest.RunCtorDoubleTests [FAIL]
13:53:28             at System.Diagnostics.Debug.Assert(Boolean condition)
13:53:28             at System.Numerics.BigIntegerCalculator.LeftShift(UInt32* value, Int32 valueLength, UInt32* target, Int32 targetLength, Int32 shift)
13:53:28             at System.Numerics.BigIntegerCalculator.Divide(UInt32* left, Int32 leftLength, UInt32* right, Int32 rightLength, UInt32* quotient, Int32 quotientLength, UInt32* remainder, Int32 remainderLength)
13:53:28             at System.Numerics.BigIntegerCalculator.Divide(UInt32[] left, UInt32[] right)
13:53:28             at System.Numerics.BigInteger.op_Division(BigInteger dividend, BigInteger divisor)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.VerifyBigIntegerUsingIdentities(BigInteger bigInteger, Boolean isZero)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.VerifyCtorDouble(Double value)
13:53:28             at System.Numerics.Tests.BigIntegerConstructorTest.RunCtorDoubleTests()
13:53:28 Finished:    System.Runtime.Tests
```