@axelheer 2015-05-04
@stephentoub thanks, I hope it's fixed now. Running `msbuild /t:BuildAndTest` locally (Windows) doesn't trigger any `Debug.Assert` statement -- any idea what's wrong with my system?