@stephentoub 2015-05-04
> Running msbuild /t:BuildAndTest locally (Windows) doesn't trigger any Debug.Assert statement -- any idea what's wrong with my system?

There was a build bug that was causing some of the Debug.Assert overloads on Windows to be empty, such that they were nops.  I thought that was already fixed in the NuGet packages, but either that's not the case (@mellinoe?) or maybe you could try clearing your package cache and rebuilding.