@stephentoub 2015-05-04
> I hope it's fixed now

Yup, appears to be.  The CI passed, including for Linux (where Debug.Assert is not broken).