@jasonwilliams200OK 2015-05-10
:+1: IMO, a millisecond save (with no side-effect) in core is worth it.