@stephentoub 2015-05-12
I left a few comments, but otherwise LGTM.

Do we already have test cases that cover all of the various inputs that would hit all of the code paths being modified?