@KrzysztofCwalina 2015-05-12
@axelheer, as to the stack overflow issues: would it make sense to have two internal implementations and switch between them depending on the size? or even provide an explicit API to select between the implementations.