@axelheer 2015-05-13
@stephentoub thanks, I'll work on them. Since I added no new functionality I just assumed we've enough tests. Is there any way to do some kind of code coverage analysis?

@KrzysztofCwalina since the operations can be called within code already using much of the limited stack size, I don't know how to handle that automatically. Furthermore, I cannot measure any performance gain in using `stackallock` instead of `fixed`. I just thought using the stack is more efficient, but now I think I was wrong. And, if we work without pointers, we'd use ordinary arrays anyway. Hence, I don't see the benefit in using `stackallock` anymore. What do you think?