@stephentoub 2015-05-13
@axelheer, you can get code coverage numbers by going into the tests directory:
```
cd corefx\src\System.Runtime.Numerics\tests
```
and running:
```
msbuild /t:rebuild /t:BuildAndTest /p:Coverage=true
```
Assuming that completes successfully, you can then view the ```index.htm``` page that's dropped into the ```corefx\bin\tests\coverage``` directory.