@axelheer 2015-05-13
@stephentoub that's a very nice coverage analysis. I've found a corner case of the integer division, which was not covered (added tests for that). We have still 12 not covered lines of code within `BigIntegerCalculator`, but those cannot be reached at the moment.

1) the carry handling of `AddSelf` and `SubtractSelf` cannot be reached (yet), since they are only used by code, which should not lead to an extra carry
2) one path of the `Compare` method for pointers is dead, but just removing it is too error prone for my taste (future code should assume that `Compare` is working...)
3) the path for 32-bits of `LeadingZeros` is dead, but again, removing it is just not correct