@stephentoub 2015-05-19
LGTM, once squashed, and other than my one open question on stackalloc vs fixed.  @KrzysztofCwalina?