@axelheer 2015-05-19
@stephentoub the build of the squashed commit has failed, but I can't see why?