@Maxwe11 2015-05-19
http://dotnet-ci.cloudapp.net/job/dotnet_corefx_windows_release_prtest/1350/console
@dotnet-bot test this please