@Maxwe11 2015-05-19
@dotnet-bot test this please