@stephentoub 2015-05-20
@dotnet-bot test this please